Rename package folders to sealedsecrets
