+++ pkg/client/clientset/versioned/typed/sealedsecrets/v1alpha1/fake/fake_sealed-secrets_client.go
@@ -3,7 +3,7 @@
 package fake
 
 import (
-	v1alpha1 "github.com/bitnami-labs/sealed-secrets/pkg/client/clientset/versioned/typed/sealed-secrets/v1alpha1"
+	v1alpha1 "github.com/bitnami-labs/sealed-secrets/pkg/client/clientset/versioned/typed/sealedsecrets/v1alpha1"
 	rest "k8s.io/client-go/rest"
 	testing "k8s.io/client-go/testing"
 )
