+++ pkg/client/clientset/versioned/typed/sealedsecrets/v1alpha1/sealed-secrets_client.go
@@ -3,7 +3,7 @@
 package v1alpha1
 
 import (
-	v1alpha1 "github.com/bitnami-labs/sealed-secrets/pkg/apis/sealed-secrets/v1alpha1"
+	v1alpha1 "github.com/bitnami-labs/sealed-secrets/pkg/apis/sealedsecrets/v1alpha1"
 	"github.com/bitnami-labs/sealed-secrets/pkg/client/clientset/versioned/scheme"
 	rest "k8s.io/client-go/rest"
 )
