+++ pkg/client/informers/externalversions/sealedsecrets/v1alpha1/sealedsecret.go
@@ -5,10 +5,10 @@ package v1alpha1
 import (
 	time "time"
 
-	sealedsecretsv1alpha1 "github.com/bitnami-labs/sealed-secrets/pkg/apis/sealed-secrets/v1alpha1"
+	sealedsecretsv1alpha1 "github.com/bitnami-labs/sealed-secrets/pkg/apis/sealedsecrets/v1alpha1"
 	versioned "github.com/bitnami-labs/sealed-secrets/pkg/client/clientset/versioned"
 	internalinterfaces "github.com/bitnami-labs/sealed-secrets/pkg/client/informers/externalversions/internalinterfaces"
-	v1alpha1 "github.com/bitnami-labs/sealed-secrets/pkg/client/listers/sealed-secrets/v1alpha1"
+	v1alpha1 "github.com/bitnami-labs/sealed-secrets/pkg/client/listers/sealedsecrets/v1alpha1"
 	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	runtime "k8s.io/apimachinery/pkg/runtime"
 	watch "k8s.io/apimachinery/pkg/watch"
